Create oddEven
This program ask user to input numbers that are positive then the program prints out the user input, the odd numbers,and the even numbers all form what the user inputted. **Disclaimer** there is a logical error i have -1 to exit the program which works but isn't pretty. solution:is to have Q be the exit instead of -1. (I think)